Refactor CSV export functionality into separate utility module
- Extracted the CSV export logic from AcfLogReader class into a new utility module app/utils/csv_export_utils.py
- Moved the export_to_csv method and all its helper methods (_prepare_csv_export, _write_csv_data, etc.) to the utility module
- Updated AcfLogReader to import and use the new utility function
- This improves code organization by separating concerns and makes the CSV export functionality reusable
